requirements.txt
pyserial
opencv-python
pyzbar
